simplify tests for @concurrent
